+++ masm-tasm/src/dosbox/jsdos/Jsdos.ts
@@ -1,22 +1,19 @@
-import * as adapted from "../emulators/main";
-import { CommandInterface } from "emulators";
+import * as adapted from "emulators";
 import * as vscode from "vscode";
 import * as api from "../api";
 import { Conf } from "../dosbox/conf";
 import * as Jszip from "jszip";
 import { createTerminal } from "./runInHost";
 import { runInWebview } from "./runInWebview";
-import { EmulatorFunction } from "emulators-ui/dist/types/js-dos";
-import { logger } from "../util/logger";
-import { JsdosShell } from "../jsdos-ci-shell/main";
 import { isNode } from "browser-or-node";
 import { loadWasmModule } from "./hooks";
 
 const fs = vscode.workspace.fs;
 
 export class Jsdos implements api.Jsdos {
+  emulators:adapted.Emulators|undefined
   public set pathPrefix(pathPrefix: string) {
-    adapted.emulators.pathPrefix = pathPrefix;
+    this.emulators=adapted.getEmulators(pathPrefix)
   }
   public conf: Conf = new Conf("");
   public jszip: Jszip = new Jszip();
@@ -25,32 +22,10 @@ export class Jsdos implements api.Jsdos {
     const dist = vscode.Uri.joinPath(context.extensionUri, "/dist/emulators/");
     this.pathPrefix = isNode ? dist.fsPath : dist.toString();
 
-    //take over path resolve method for file in extension has different schema
-    // - in Browser has a schema of http
-    // - in Machine has a schema of file
-    // adaptedEmulatorsImpl.pathResolver = function (filename: string) {
-    //   const uri = vscode.Uri.joinPath(dist, filename);
-    //   return dist.scheme === "file" ? uri.fsPath : uri.toString(true);
-    // };
-
-    //take over HTTP request for running as web extension
-    const request: adapted.ReadContent = async function (
-      url: string,
-      options: any
-    ): Promise<string | ArrayBuffer | undefined> {
-      const uri = vscode.Uri.parse(url);
-      const fileArr = await fs.readFile(uri);
-      console.log(url, options);
-      if (uri.path.includes(context.extensionUri.path)) {
-        if (options.responseType === "arraybuffer") {
-          return fileArr;
-        }
-        return new TextDecoder().decode(fileArr);
-      }
-      return undefined;
-    };
-    adapted.setReadContent(request);
-    adapted.setLoadWasmModule(loadWasmModule);
+    const oldhttp=adapted.platform.current.httpRequest
+    adapted.platform.current.httpRequest=function (url: string, options: adapted.XhrOptions) {
+      return oldhttp(url,options)
+    }
   }
 
   async setBundle(
@@ -97,18 +72,19 @@ export class Jsdos implements api.Jsdos {
   public async runInHost(
     bundle?: vscode.Uri | null | undefined,
     useWorker?: boolean
-  ): Promise<api.CI> {
+  ): Promise<adapted.CommandInterface> {
     const ci = await this._runInHost(bundle, useWorker);
-    const shell = new JsdosShell(ci as CommandInterface);
-    (ci as api.CI).shell = shell;
-    (ci as api.CI).terminal = () => createTerminal(shell);
-    return ci as api.CI;
+    const shell = new adapted.utils.Shell(ci);
+    return ci;
   }
   private async _runInHost(
     bundle?: vscode.Uri | null | undefined,
     useWorker?: boolean
   ): Promise<adapted.CommandInterface> {
-    let func: EmulatorFunction = (process as any).browser
+    if (typeof this.emulators ==="undefined"){
+      throw new Error("emulators not initialized")
+    }
+    let func:keyof adapted.Emulators = (process as any).browser
       ? "dosboxWorker"
       : "dosboxDirect";
     if (useWorker === true) {
@@ -116,18 +92,19 @@ export class Jsdos implements api.Jsdos {
     } else if (useWorker === false) {
       func = "dosboxDirect";
     }
+
     if (bundle === undefined) {
       const bundleData = await this.getBundleData();
-      const ci = await adapted.emulators[func](bundleData);
+      const ci = await this.emulators[func](bundleData);
       return ci;
     } else if (bundle === null) {
       const bundleData = await new Jszip()
         .file(".jsdos/dosbox.conf", "")
         .generateAsync({ type: "uint8array" });
-      return adapted.emulators[func](bundleData);
+      return this.emulators[func](bundleData);
     } else if (bundle.scheme === "file") {
       const bundleData = await fs.readFile(bundle);
-      return adapted.emulators[func](bundleData);
+      return this.emulators[func](bundleData);
     }
     throw new Error(
       "bundle uri is not a uri with schema file or undefined/null"