+++ masm-tasm/src/dosbox/test/suite/util.ts
@@ -1,3 +1,0 @@
-export function randomString() {
-  return Math.floor(Math.random() * 100000).toString();
-}