+++ masm-tasm/webpack.config.js
@@ -108,9 +108,49 @@ const webExtensionConfig = {
     devtool: 'nosources-source-map', // create a source map that points to the original source file
 };
 
-module.exports = [webExtensionConfig];
+const webviewConfig = {
+  entry: path.resolve(__dirname, "src/dosbox/webview/index.ts"), // the entry point 📖 -> https://webpack.js.org/configuration/entry-context/
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "index.js",
+    devtoolModuleFilenameTemplate: "../[resource-path]",
+  },
+  externals: [],
+  devtool: "source-map",
+  resolve: {
+    // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
+    extensions: [".ts", ".js"],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        exclude: /node_modules/,
+        use: [
+          {
+            // configure TypeScript loader:
+            // * enable sources maps for end-to-end source maps
+            loader: "ts-loader",
+            options: {
+              compilerOptions: {
+                sourceMap: true,
+              },
+            },
+          },
+        ],
+      },
+    ],
+  },
+  optimization: {
+    minimize: process.argv.includes("--mode=production"),
+  },
+  stats: {
+    warnings: false,
+  },
+};
 
 module.exports = [
     config,
-    webExtensionConfig
+    webExtensionConfig,
+    webviewConfig
 ]